Added protocol addon functionality
When the user is logging in, they may or may not enter the server name with
a ws:// or a wss:// preceding the server name. With a simple regexular
expression, this code will now add the correct protocol to the beginning of
the server name string before the socket is created.