added doc and error handling
Added documentation for the TransformerMessage, and added documentation for the init and sendMessage methods. Upon writing the documentation for sendMessage, I realized that a potential error ould arise if the server had not yet been properly initialized, so I created a custom error NotInitializedException. Now thinking about it, there will probably be more errors in the future, like if the socket was not successfully created.