added getMessage
Now, the socket will listen for messages from the server and will stick
them into an array. This is all done by the onmessage event of Websocket.
Then, when the main game code is ready to accept/receive the server message,
it simply needs to call .getMessage, which gets the gets the first message
in the messages queue (in actuality, this is the last element of the array),
and returns it. This is done easily by using Array.prototype.shift