fixed double click bug, disables button after click
Before, there was  bug where, when attempting to log in, the user would be forced to rapidly double click the "Log In" button in order for the screen to change successfully. Now, using something that I read on StackOverflow, the button only needs to be clicked a single time. This utilizes $rootScope and $rootScope's $apply method.

Now, on the login page, after the user has entered their credentials and they have hit the "Log In" button, this button becomes instantly disabled so that multiple requests aren't sent to the server.