now uses $interval to update chat and users
Every two seconds, this will now grab a message from the server. If the message is a "userstatus" message, it will add this user to the user list (and will later store the information about the user). If it is a chat message it will add it the chatMessages list. In the future, however, there needs to be (at least) some conditional checking for the "userstatus" messages because we don't want the list to display offline users.

However, this is extremely difficult to test because the server is not accepting my Login request.