MAJOR FIX now, the API correctly sends messages
Before, when the API was sending a request to the server, it would not include the .command of the request. Why? JSON.stringify doesn't include the inherited properties, which in our case, is .command from Message. After a quick search, I found on SO a simple function for creating a new object with all the properties needed. Now, this function is performed on the message before it is converted to JSON and before that JSON is sent to the server.