correctly parses messages from server, new onReceive argument on sendMessage
I don't know how this slipped by before, but now the code will JSON.parse each incoming message before storing it in the incomingMessages array.

The onReceieve argument of sendMessage now allows the game code to set a function to be run upon receiving a response message to the message they just sent. For example, this would be helpful when sending a LoginMessage, so you can check the welcome message to make sure everything went smoothly.