changed buttons that send info to server to type "submit", added event listener for "enter" key
Normally in the Log In page you'd have to move your mouse all the way across the screen and then push your finger down on the mouse to click on the "Log In" button and send your login credentials; that's too much work. Now, to log in, all you have to do is hit the enter key. Now your fingers never have to leave the comfort of your keyboard.

Changed the text input part of the chat form to a text area, like chat.stackexchange.com uses.

In the sendMessage function of the lobby controller, a small bit of code has been added to read the user's keystrokes as they are entering the message, and then to send the message if the user hits the "enter" key. This has to be done to allow "enter" to be used to send the message because just changing the input button type to "submit" will not work: hitting enter will just move you down to the next line in the text area.