removed getMessage, added setMessageListener, KISSed sendMessage
setMessageListener is a new method that is aiming to replace the message queue system. When you call it, you can pass in a set of types. These types describe the kinds of messages that the listener should respond to if received. For example, if you did:

CardshifterServerAPI.setMessageListener(function(message) {
     ...
}, ["chat", "userstatus"]);

Then, every time a message was received and it's .command is either "chat" or "userstatus", then the function passed in will be called and the received message will be passed in.