now conforms to the new API, setup a command map.
This command map is an object where the keys are the .command s of messages, and the values are functions. When a message is received, it's command is used to find the function to call that will act on the newly received message. This is definitely much more structured and cleaner than the old lobby controller.