setup a counter for how many cards are being used, a box for card information, and a list of each card.
Each list item will contain the name of the card and two buttons: a + and a - for choosing and deselecting a card. The card name is a link/button that will display all the information about that card at the top of the document.

The information display (right now) is extremely ugly.