sets up card maximum values, create showDetails function, created getTotalSelected function, created increment and decrement functions
When the first message is received from the server describing all the cards, a new property is set to each card called .max. This will either hold the card's personal maximum hold limit, or it will hold the default maximum hold limit.

The showDetails function will display all the information about a card at the top of the document. Right now, it is extremely ugly.

The getTotalSelected function will sum up all the values in $scope.cardChoice (which is all the cards and how many times the user has selected them) and return the total amount

The increment and decrement functions correspond to the + and - buttons of each card. The increment will not go over the maximum allowed cards and the maximum allowed cards for that specific card, and the decrement function will not let the card amount go below 0 for that specific card.