added feature for saving created decks which are saved in LocalStorage
The function $scope.saveDeck is used to create a saved deck. This deck will not save if (1) the user has not selected enough cards, (2) the user has not  yet entered a name for the deck, or (3) there is already another deck with that name. This is run from the deck creation form.

The function $scope.switchDeck is used for changing cards to that of another deck. This is run from the ng-repeat list in the HTML of all the decks.

The function updateSavedDecks will update the $scope.savedDecks variable when a new deck has been added. This is called from $scope.saveDeck after the deck has been saved. Calling this function allows the HTML to update without an error.

The functino deckExists takes a single name parameter and checks all the saved decks to see if there is another one with the same name.