added delete buttons
For each deck in the list of saved decks, there is a little button for deleting a deck.

IDEA: In the case that this is not good, here is an idea for rearrangement: instead of having clicking on a deck name load it up, it will just set it to "selected". Then, the user can use two buttons: "delete" and "load". The "delete" button will delete the selected deck, and the "load" button will load the selected deck. Perhaps, upon selection, a special CSS class is applied to it so it can stand out.