added deck deletion, repurposed deckExists, change storing method
There is now a function $scope.deleteDeck that takes the name of a deck, searches for the deck by name in the current mod category, and removes it.

The function deckExists used to return true if the deck exists, and false if it did not. However, now it will return the index at which the deck was found, or it will return false. This allows for the $scope.deleteDeck function to not have to create a loop just to get the index of the deck. And, since numbers are treated as a true-y value in JavaScript, this function will behave the same way in $scope.saveDeck

Before, decks would just be stored in LocalStorage with a name and a string. However, I soon realized a problem: you would see decks from ALL other mods in the deck creator, which can only lead to trouble. Now, the storage is formatted like this: {decks: {"mod1": [...decks...], "mod2": [...decks...]}}. That way, decks are saved with whichever mod they correspond to.