added "Start game" and "Go back to lobby" buttons
The "Start game" button is only visible when the user is transfering to a game, but is in the middle of deck creation first. Since there is no game screen yet, this does nothing.

The "Go back to lobby" button is only visible when the user is configuring their decks of cards after being in the lobby, and are not entering a game.