started working on handling the moving of cards.
For later: The server gets an error at some point in the running of this client and I think I know why: when this client is passing all the chosen cards to the server, it all passes cards where there are 0 of them. The server may not want the client to specify cards with 0 of them.