added buttons
The first button is a submit button with a changing value: this value changes depending on the current action that is allowed by the server (may have to resdesign this, based on what the server sends).

The second button is a cancel button that will stop the current action and is only shown if the user is attempting to do an action.