added list of repeating action buttons
Rather than having two buttons (one for an action and one for cancel which was extremely inflexible because there are multiple possible actions at once, and cancel isn't always possible), the code now uses `ng-repeat` to repeat as many buttons as needed to show all the user all their possible actions.

This is coming along quite well right now. Getting the action to the from the user should be fairly easy. However, creating the interface and getting the correct cards to show could be a bit difficult.