now listens for resetActions and for usable
Upon receiving a resetActions, the code will clear the $scope.actions list.

Upon receiving a usable, the code will add a single action to the $scope.actions list.

Upon receiving either of these messages, the $scope.actions list will be changed which will cause the HTML to change, either deleting all the buttons or adding a button for completing an action.