refactored storePlayerInfo, now adds to $scope.playerInfos
Before, depending on whether the player info message was talking about this user or if it was talking about the opponent, the code would have a conditional to set all the correct properties of currentUser.game.<player type>.<etc>.

However, this had a lot of repeat code because the only difference between the code in the two conditionals was where it was storing the data. Now, instead of all that, the code will store the player type (either "playerInfo" or "oppInfo") in a local variable and use that to access the correct property of currentUser.game.

This method eliminates all repeat code. And, at the very end of this function, all the information about that player is stored in $scope.playerInfos so the HTML can then load the information into a visual on the page.