created local playerInfos
Before, to store information about each player, this code was using currentPlayer. However, this was a poor idea because there is no other scene that needs that information so there is no point in storing it there.

A few parts of this code has been rewritten to adjust to the new place for storing and reading the information.