now stores zones in playerInfos.<user>.zones.<zone_name>, removed storeCard
Before, upon receiving a ZoneMessage, the code would store the zone message by it's ID in a zones object. However, this didn't really help the code at all because it wasn't stored in a place that would make sense to the code: each timethat  the code would want to use a zone, the code would have to perform a bunch of conditional checks.

Now, by storing all the zone information in each player's information, the code can easily access all the sections of cards that a player has by simply reading the player's .zones.

<hr>

Before, there was a method called `storeCard` that would store all necessary card information in an array of cards. Yes, while this should have been put in the "Hand" zone of a player, there was another issue: when the zone message about a player's hand is received, it already contains all the cards that that player has; there is no need for the code to add these manually.

There is another plus to removing this: there are no longer any fixed-length hand sizes.