setZone now check to make sure the owner is correct, now skips "Cards"
Before, the code would just stick the Zone information in both of the players. However, because of the order in which the messages were sent, this worked (I think). Now, before storing the zone information inside of a player, the function checks to make sure that the owner of the zone is the owner that is the current object of the iteration.

<hr>

According to https://github.com/Cardshifter/Cardshifter/blob/master/Documentation/2-%20Card%20Library%20-%20Effects.md , the "Cards" type zone is "Not currently used as it is too meta." So now, the setZone function will skip all messages of this type.