dynamically loads player information
This HTML takes advantage of the player information that is stored in $scope.playerInfos. For each, player there are three sections created: the user's name, the user's properties (health, mana, etc), and the user's KNOWN card

The user's name is displayed with an underline in an <h1> tag

The user's properties are stored in an ng-repeated list, where the property name is bold and it's value is a ": " away.

The user's cards are displayed with the name of the zone in an <h3> tag and then a ng-repeated list of the cards name. In the future, the name will be a link for selecting the card.