added optional link for each card entity
When the user is attempting to perform an action, the action can only work on certain targets.

Now, for each card being displayed, there are two displays: a link, and a normal paragraph. If that card is an available target, then the link will be displayed. By clicking this link, the card will be selected/deselected. However, if the card is not an available target, then the simple paragraph text will be displayed.