"clicker" mouse display now shows for links
Before, when going to click on one of the "links", the mouse symbol that looks like a capital "i" would be shown. This, to the user, does not look like something that should be clicked.

To fix this, a blank "href" attribute was added to these link tags so now the mouse will form the small hand that is used looks like it is trying to click something.