servers are stored in an object
Before, all the available servers were hardcoded into the select in the login form of the login page. This made adding new servers a bit more difficult, because an entire new element would have to be added each time, which only leads to copy paste coding.

Now, all the servers are stored in $scope.servers, which is an object that is mapped [server name: server address]. Then, through use of ng-repeat, a new <option> is created for each server under the select.