now server is stored with information like user count and latency
Before, servers were stored in an object like this: [name: address]. Of course, this does not include enough information for the user. Is the server online? What about the users? How many of those are these? And the latency? I'm I going to be stuck trying to send a chat message?

Now, the servers are stored in an object like this:

name:
    address
    isOnline
    userCount
    latency