added updateStats function
The updateStats function so far initializes the API and gets a latency count by storing the time before the init call and getting the time after, and the code can easily find out if the server is online if the connection is made at all.

This currently does not work and is being tested.