fixed typo in server map, added checking in updateStats so there is no checking for "Other..."
Now the code sort of works. It will only display information for stats.zomis.net, but the latency is blown out of proportion: I was reaching 2000+ ms. I think the reason for this is because the server is being queried to quickly, so by the time init's third parameter runs, it is using an old local `now` variable.

Another issue: this will not recognize a local host server.