added refresh button
Instead of checking the server stats every 2 seconds, there is now a refresh button. There are still some bugs to be worked out, for example, if the server does not exist (local host), then the code dies. And, if the server does exist, the code may sit waiting for a response indefinitely, even though one may never be received. And, this refersh button allows for an easy attempt of a DoS on the server.

Maybe this idea should be put off for a little while.