added handling for UpdateMessages, added findUser(by id)
The findUser by ID function is a function that will either return playerInfos.user or playerInfos.opponent based on the ID that was passed in. In case it is useful later, the function will return null if the ID does not pertain to either user.

Now, upon receiving an UpdateMessage, the code does this simply: finds the user with UpdateMessage.id using the new findUser function, and set's it's UpdateMessage.key property to UpdateMessage.value.