changed addUsableAction function
Now, this function will only add the action to the $scope.actions if there is no other action with the same name in the array. The reason for this is because, the first exception being mulligan, there may be multiple UsableActionMessages sent for each card.

This currently does not work.