fixed typo in conditional of addUsableAction
In addUsableAction, there is a loop that checks all the currently stored actions' names to make sure that a new button isn't created if there is already a button with the same name on the screen. However, in the conditional, this was being checked:

actions[i].action === actions.action

The problem? actionS.action. This is trying to access the action property of the actions *array*. The code should be checking the variable "action", which is the name of the argument being passed into this function.