added local notDuplicate variable to addUsableAction
If the conditional inside the loop passed, then the action would be added to the actions array. While this itself is fine, the looping was a problem: if there was a non-duplicate at any point in the array, the action would be added.

To fix this, if there is a duplicate, the code will set notDuplicate to false and break from the loop. Outside the loop, there is another conditional: if notDuplicate is true, THEN added the action to the array.