$scope.performAction now sends selected card IDs
Before, when creating the UseAbilityMessage, $scope.perform action would pass in $scope.targets. However, that is a problem because that is passing in all of the available targets to the server. So now, this function grabs the IDs of all the cards in $scope.selected, shoves them into an array, and then passes that array along in the message.