added handling for card actions in $scope.performAction
As mentioned before, some actions are performed by the card rather than the player. So now, in $scope.performAction, a conditional was set up to check if the thing that is supposed to perform the action is a player or a card.

Why? For an action performed by a card, the targets array is [0] and the ID is the card ID, and for an action performed by the player, the targets array is an array of all the cards that are targets of this action and the ID is the user's ID.