reduced repeating in storing and fetching of previous login data
Before, when either storing or fetching data from local storage, there would be four practically identical lines to complete the action, which is not very pretty.

To fix this, a map/object was created called loginStorageMap. In this map, the keys are the name of the place in local storage where this data is being stored (for example, CARDSHIFTER_LAST_NAME). The values are the $scope property that will later be stored at this memory location and right now will be loaded with whatever is in that memory location (for example, username, for $scope["username"])

Now, whenever either storing or fetching is done with memory, a simple for/in loop is created to loop through this map and interact with memory based on everything described in the map.