actions performed by cards get buttons right next to the card
Before, any and all actions would be put up at the top of the screen with the rest of the actions. This led to great confusions both on the UX side and the code side because some actions like Play are sent multiple times for each card that can perform this action. Therefore, the server would not like it if a RequestTargetsMessage was sent for a message like that, because the UsableActionMessage would already contain the targets.

Now, if a UsableActionMessage is received and the ID is that of a card, then a button with the name of the action will be placed right next to card name. If the ID is that of a player, then the action will be placed up at the top.