displays card model on click
Before, when a card name link was clicked, a list of all the card data would be displayed at the top. However, now that there is a card model directive, upon clicking the card name link, a card model directive is loaded which displays all the information about the card in card form.