displays win or lose message
Now, when a player is eliminated, the current user is shown a message either telling them that they won or that they lost. Either way, after that, the player is navigated back to the lobby.

Right now, the message is displayed in the form of an alert (AH!). Later, this will be replaced with ui-bootstrap magic.