added clickable radio buttons for mods
At the bottom of the screen is the mod selection and on the side of the screen is the player list. Originally, in order to select one of the elements in either of these lists, you'd have to click that tiny radio button. However, now you can just click the label associated with it to select it.