removed "cancel" button from game over pop-up
Before, in the game over pop-up, there were two buttons: "OK" and "Cancel". The "Cancel" button doesn't make any sense because the pop-up is merely a declarative statment; there is nothing for the user to deny. Also, the button's code logged unnecessary information to the console and left the user in a stuck state in the game board screen.