only tries to display "from" in a chat message if it exists, cleaned up chat error grammar
Before, the error messages that are shown in chat from not selecting an opponent and/or a game type before using deck builder or staring a new game, the "from" field of the message would try to be displayed, so a funky space and a : would be shown in it's place. This is no longer there.