Initial commit; using react-slingshot
Re-laying out the repo with a design based off the one given from react-slingshot. Under each src/scene will be all the code for each scene.

This is a temporary template; not everything will be needed (for example, I don't think we'll need /containers/). In each subfolder you'll find, for example, actions.js. This does not mean to put all the actions in there; this a file to group all the actions in that directory and export them.

The index.js file of each scene is meant to serve the same purpose; it gathers together all of it's assets and exports them. I am unsure of how style sheets are incorporated so, for now, there will be one for each scene.

The shared folder is meant for shared services. I'm not sure if this will be needed either.

The store and tools folders come from react-slingshot.