remove export default const = syntax
Apparently this is invalid syntax. This is news to me, however this
does mean that the App component we have here has no name now. :(

Note that the following still works:

````
export const App = () => <div />

const App = () => <div />
export default App
````

Although in the case of the latter, it appears that Babel does not
assign a name to that component - or, at least, eslint complains about
the react/display-name rule.